LASB-4333 - Fix repository Spring auto-detect
Suffix repository class with Impl for Spring auto-detect
