+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/config/AsyncConfig.java
@@ -17,7 +17,7 @@ public class AsyncConfig implements AsyncConfigurer {
 
     @Value("${aws.region}")
     private String awsRegion;
-    
+
     @Bean
     public S3Client s3Client() {
         return S3Client.builder().region(Region.of(awsRegion)).build();
