--- maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/repository/StatusRepository.java
@@ -1,9 +1,0 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.repository;
-
-import java.util.List;
-
-public interface StatusRepository<ID> {
-
-    List<ID> findUnprocessedIds();
-
-}
