+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/repository/TrialDataStatusRepository.java
@@ -1,20 +1,8 @@
 package uk.gov.justice.laa.maat.scheduled.tasks.repository;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
-@Repository
-@RequiredArgsConstructor
-public class TrialDataStatusRepository implements StatusRepository<Integer> {
+public interface TrialDataStatusRepository {
 
-    private final JdbcTemplate jdbcTemplate;
-
-    @Override
-    public List<Integer> findUnprocessedIds() {
-        String sql = "SELECT ID FROM HUB_TRIAL_DATA WHERE STATUS = 'UNPROCESSED'";
-        return jdbcTemplate.queryForList(sql, Integer.class);
-    }
+    List<Integer> findUnprocessedIds();
 }
