+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/repository/XhibitTrialDataRepository.java
@@ -3,5 +3,5 @@ package uk.gov.justice.laa.maat.scheduled.tasks.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uk.gov.justice.laa.maat.scheduled.tasks.entity.XhibitTrialDataEntity;
 
-public interface XhibitTrialDataRepository extends JpaRepository<XhibitTrialDataEntity, Integer>, StatusRepository<Integer> {
+public interface XhibitTrialDataRepository extends JpaRepository<XhibitTrialDataEntity, Integer>, TrialDataStatusRepository {
 }
